refactor: Cookie 객체 추가
